Add synchronized collection access

+++ server/src/main/kotlin/labs/commands/ClearCommand.kt
@@ -12,7 +12,7 @@ import labs.utility.CollectionManager
  */
 class ClearCommand(private val collectionManager: CollectionManager) :
     Command("clear", ": очистить коллекцию.") {
-    override fun execute(request: Request): Response {
+    override suspend fun execute(request: Request): Response {
         if (request.args.isNotEmpty()) {
             return Response(ResponseStatus.WRONG_ARGUMENTS, "Для этой команды не требуются аргументы!")
         }
@@ -26,6 +26,7 @@ class ClearCommand(private val collectionManager: CollectionManager) :
                 .filter { it!!.creatorLogin == request.user!!.login }
                 .map { it!!.id }
                 .toList()
+        if (ids.isEmpty()) return Response(ResponseStatus.WARNING, "В коллекции нет Ваших элементов!")
         if (DatabaseConnector.databaseManager.deleteAllObjects(ids, request.user!!)) {
             collectionManager.removeElements(ids)
             return Response(ResponseStatus.OK, "Ваши элементы успешно удалены.")
